refatorei o código para não precisar mais do CalendarDto
