+++ src/main/java/br/edu/ifpb/dac/sape/business/service/SchedulingService.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,13 +13,11 @@ import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.ExampleMatcher.StringMatcher;
 import org.springframework.stereotype.Service;
 
-import br.edu.ifpb.dac.sape.model.entity.Place;
 import br.edu.ifpb.dac.sape.model.entity.Scheduling;
 import br.edu.ifpb.dac.sape.model.entity.User;
 import br.edu.ifpb.dac.sape.model.enums.IsPresent;
 import br.edu.ifpb.dac.sape.model.enums.StatusScheduling;
 import br.edu.ifpb.dac.sape.model.repository.SchedulingRepository;
-import br.edu.ifpb.dac.sape.presentation.dto.CalendarDTO;
 import br.edu.ifpb.dac.sape.presentation.exception.MissingFieldException;
 import br.edu.ifpb.dac.sape.presentation.exception.ObjectNotFoundException;
 import br.edu.ifpb.dac.sape.util.EmailSender;
@@ -58,20 +55,6 @@ public class SchedulingService {
 
 		return schedulingsBeginingToday(list);
 	}
-	//metodo criado para o get scheduling do calendário no formato que a API do calendário precisa
-	public List<CalendarDTO> findAllEvents(){
-		List<Scheduling> list = new ArrayList<>();
-		list = this.findAll();
-		
-		List<CalendarDTO> listCalendar = new ArrayList<>();
-		for(Scheduling s: list) {
-			listCalendar.add(new CalendarDTO(s.getId(), s.getSport().getName(), s.getPlace().getName(),
-					s.getScheduledDate() +"T"+ s.getScheduledStartTime()+":00",
-					s.getScheduledDate() +"T"+ s.getScheduledFinishTime()+":00"));
-		}
-		return listCalendar;
-	}
-	
 	
 	public List<Scheduling> findAllByPlaceId(Integer id) {
 		return schedulingRepository.findAllByPlaceId(id);
@@ -132,7 +115,7 @@ public class SchedulingService {
 	    	Set<User> responsibles = scheduling.getPlace().getResponsibles();
 	    	emailSender.notifyPlaceResponsibles(placeId,responsibles,scheduling);
 	    }
-		
+	 
 		return schedulingRepository.save(scheduling);
 	}
